renaming file to remove duplicate name
